+++ src/App/backend/src/Altinn.App.Core/Features/Signing/Services/ISigningCallToActionService.cs
@@ -1,7 +1,6 @@
 using Altinn.App.Core.Features.Correspondence.Models;
 using Altinn.App.Core.Internal.Process.Elements.AltinnExtensionProperties;
 using Altinn.App.Core.Models;
-using Altinn.Platform.Register.Models;
 
 namespace Altinn.App.Core.Features.Signing.Services;