+++ src/Runtime/localtest/src/Services/Register/Implementation/PartiesWrapper.cs
@@ -1,5 +1,4 @@
-using Altinn.Platform.Register.Enums;
-using Altinn.Platform.Register.Models;
+using Altinn.Register.Contracts.V1;
 
 using LocalTest.Services.Register.Interface;
 using LocalTest.Services.TestData;
@@ -26,7 +25,7 @@ namespace LocalTest.Services.Register.Implementation
         public async Task<Party?> GetParty(int partyId)
         {
             var data = await _testDataService.GetTestData();
-            
+
             Party? party = data.Register.Party.TryGetValue(partyId.ToString()!, out var value) ? value : null;
 
             await AddPersonOrOrganization(party);
@@ -83,10 +82,10 @@ namespace LocalTest.Services.Register.Implementation
             switch (party.PartyTypeName)
             {
                 case PartyType.Person:
-                    party.Person = await _personService.GetPerson(party.SSN);
+                    party.Person = await _personService.GetPerson(party.SSN!);
                     break;
                 case PartyType.Organisation:
-                    party.Organization = await _organizationService.GetOrganization(party.OrgNumber);
+                    party.Organization = await _organizationService.GetOrganization(party.OrgNumber!);
                     break;
             }
         }